requirements.txt
CherryPy==3.2.2
Django==1.4
South==0.7.5
amqplib==1.0.2
anyjson==0.3.1
argparse==1.2.1
bambu-tools==1.8
celery==2.5.3
certifi==0.0.8
chardet==1.0.1
colorama==0.2.4
django-reversion==1.6
django-celery==2.5.5
django-dajax==0.8.4
django-dajaxice==0.2
django-durationfield==0.3.3
django-extensions==0.8
django-mockups==0.4.8
django-notes==0.2.2
django-picklefield==0.2.1
django-pyvows==0.6.1
gevent==0.13.7
googleappsauth==1.03
greenlet==0.3.4
gunicorn==0.14.3
httplib2==0.7.4
kombu==2.1.8
lxml==2.3.4
oauthlib==0.1.3
pip==1.1
pyVows==1.0.0
pyasn1==0.1.3
python-dateutil==1.5
requests==0.12.1
rsa==3.0.1
setuptools==0.6c11
wsgiref==0.1.2
yolk==0.4.3
django-reversion-compare
diff-match-patch
django-choices==1.1.9
